+++ drake/systems/controllers/gravity_compensator.h
