+++ drake/systems/controllers/test/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_executable(gravity_compensator_test gravity_compensator_test.cc)
+target_link_libraries(gravity_compensator_test
+    drakeSystemFramework
+    drakeSystem2Controllers
+    drakeRBSystem ${GTEST_BOTH_LIBRARIES})
+add_test(NAME gravity_compensator_test COMMAND gravity_compensator_test)
